Added som tests
Most of the tests has been made.
The findNextJob method might be broken. Seems like the method keeps on
going with no end...

+++ BenchmarkSystem/BenchmarkSystem/Jobs/Job.cs
@@ -42,7 +42,7 @@ namespace Jobs
 
 		public readonly JobType type;
 
-		public int numberOfDelays = 0;
+		public short numberOfDelays = 0;
 
 		/// <summary>
 		/// Creates a new job and adds it to the database. After adding to the database, it assigns a ID to itself.
